削除機能更新 DAOとBeanをまとめた
