+++ src/main/webapp/views/manageDeleteOut.jsp
@@ -1,0 +1,11 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"%>
+<%@include file="../header.html"%>
+
+<p>削除しました</p>
+
+<p>
+	<button type=button
+		onclick="location.href='/ShoppingSite/views/manageStart.jsp'">商品管理ページへ</button>
+</p>
+
+<%@include file="../footer.html"%>
